+++ packages/geodata/CHANGELOG.md
@@ -14,6 +14,9 @@ adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 - Renamed "Turkey" to "Türkiye" in all maps.
 - Saint Kitts and Nevis is now specified as belonging North America in `countries2` data file.
 
+### Fixed
+- Västernorrland and Västmanland had their places switched in the maps of Sweden.
+
 
 ## [5.0.3] - 2022-03-31